# backend/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Blockchain Gateway Backend API
  version: 3.0.0
  description: |
    Backend API for chaincode lifecycle orchestration with RBAC.
    
    This API provides comprehensive management of:
    - User authentication and authorization with JWT tokens
    - Role-based access control (ADMIN, ORG_ADMIN, USER, VIEWER)
    - Chaincode lifecycle management (upload, validate, approve, deploy)
    - Deployment orchestration with workflow management
    - Certificate synchronization with Fabric CA
    - Audit logging for all operations
    
    ## Authentication
    
    The API uses JWT tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Roles and Permissions
    
    - **ADMIN**: Full system access, can manage users, approve chaincodes, configure system
    - **ORG_ADMIN**: Organization administrator, can upload/deploy chaincodes, view users
    - **USER**: Regular user, can invoke/query chaincodes, manage assets
    - **VIEWER**: Read-only access, can query chaincodes and view assets
    
  contact:
    name: Blockchain Gateway Team
    email: support@blockchain.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000/api/v1
    description: Development server
  - url: https://api.blockchain.com/api/v1
    description: Production server
...
paths:
...
  /chaincode/upload:
    post:
      tags:
        - Chaincode Management
      summary: Upload chaincode
      description: Upload a new chaincode (Org Admin and above)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChaincodeUploadRequest'
      responses:
        '201':
          description: Chaincode uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChaincodeResponse'
        '400':
          description: Invalid chaincode data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /chaincode:
    get:
      tags:
        - Chaincode Management
      summary: Get chaincodes list
      description: Get list of chaincodes with optional filters
      security:
        - bearerAuth: []
      parameters:
        - name: skip
          in: query
          description: Number of records to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Number of records to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: status
          in: query
          description: Filter by status
          required: false
          schema:
            type: string
            enum: [uploaded, validated, approved, rejected, deployed, active, deprecated]
        - name: uploaded_by
          in: query
          description: Filter by uploader
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Chaincodes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChaincodeListResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /chaincode/{chaincode_id}:
    get:
      tags:
        - Chaincode Management
      summary: Get chaincode by ID
      description: Get chaincode information by ID
      security:
        - bearerAuth: []
      parameters:
        - name: chaincode_id
          in: path
          required: true
          description: Chaincode ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Chaincode retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChaincodeResponse'
        '404':
          description: Chaincode not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Chaincode Management
      summary: Update chaincode
      description: Update chaincode information (Org Admin and above)
      security:
        - bearerAuth: []
      parameters:
        - name: chaincode_id
          in: path
          required: true
          description: Chaincode ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChaincodeUpdateRequest'
      responses:
        '200':
          description: Chaincode updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChaincodeResponse'
        '400':
          description: Invalid chaincode data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Chaincode not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /chaincode/{chaincode_id}/approve:
    put:
      tags:
        - Chaincode Management
      summary: Approve chaincode
      description: Approve a chaincode (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: chaincode_id
          in: path
          required: true
          description: Chaincode ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Chaincode approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChaincodeResponse'
        '404':
          description: Chaincode not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /chaincode/{chaincode_id}/reject:
    put:
      tags:
        - Chaincode Management
      summary: Reject chaincode
      description: Reject a chaincode with reason (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: chaincode_id
          in: path
          required: true
          description: Chaincode ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - rejection_reason
              properties:
                rejection_reason:
                  type: string
                  example: "Code quality issues"
      responses:
        '200':
          description: Chaincode rejected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChaincodeResponse'
        '404':
          description: Chaincode not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Deployment Management endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        detail:
          type: string
          example: "Detailed error information"

    # Authentication schemas
...
    ChaincodeUploadRequest:
      type: object
      required:
        - name
        - version
        - source_code
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "basic"
        version:
          type: string
          minLength: 1
          maxLength: 20
          example: "1.0"
        source_code:
          type: string
          example: "package main\n\nimport (\n\t\"fmt\"\n\t\"github.com/hyperledger/fabric-contract-api-go/contractapi\"\n)\n\nfunc main() {\n\t// Chaincode implementation\n}"
        description:
          type: string
          example: "Basic asset management chaincode"
        language:
          type: string
          enum: [golang, java, javascript, typescript]
          default: "golang"
          example: "golang"
    ChaincodeUpdateRequest:
      type: object
      properties:
        description:
          type: string
          example: "Updated description"
        metadata:
          type: object
          additionalProperties: true
          example:
            author: "John Doe"
            version_notes: "Bug fixes"
    ChaincodeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "basic"
        version:
          type: string
          example: "1.0"
        source_code:
          type: string
          example: "package main\n\n..."
        description:
          type: string
          example: "Basic asset management chaincode"
        language:
          type: string
          example: "golang"
        status:
          type: string
          enum: [uploaded, validated, approved, rejected, deployed, active, deprecated]
          example: "uploaded"
        uploaded_by:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        approved_by:
          type: string
          format: uuid
          nullable: true
          example: "123e4567-e89b-12d3-a456-426614174000"
        approval_date:
          type: string
          format: date-time
          nullable: true
          example: "2023-12-01T10:30:00Z"
        rejection_reason:
          type: string
          nullable: true
          example: "Code quality issues"
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        created_at:
          type: string
          format: date-time
          example: "2023-12-01T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-12-01T10:30:00Z"
    ChaincodeListResponse:
      type: object
      properties:
        chaincodes:
          type: array
          items:
            $ref: '#/components/schemas/ChaincodeResponse'
        total:
          type: integer
          example: 10
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 100

    # Deployment Management schemas
...